api_crud_bike_store/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlmodel==0.0.22
pyodbc==5.2.0
pydantic
azure-identity
python-dotenv